style(backend): rename and merge some files
Merge term_handler and socket_handler into socket_handlers.rs, move page handlers into page_handlers.rs
