+++ src/backend/src/main.rs
@@ -5,10 +5,10 @@ use simple_logger::SimpleLogger;
 use warp::Filter;
 
 mod config;
+mod page_handlers;
 mod shared;
-mod sockets;
+mod socket_handlers;
 mod systemdata;
-mod terminal;
 
 #[allow(clippy::too_many_lines)]
 fn main() {
@@ -98,11 +98,16 @@ fn main() {
 
             let terminal_route = warp::path!("ws" / "term")
                 .and(warp::ws())
-                .map(|ws: warp::ws::Ws| ws.on_upgrade(terminal::term_handler));
+                .map(|ws: warp::ws::Ws| ws.on_upgrade(socket_handlers::term_handler));
 
             let socket_route = warp::path("ws")
                 .and(warp::ws())
-                .map(|ws: warp::ws::Ws| ws.on_upgrade(sockets::socket_handler));
+                .map(|ws: warp::ws::Ws| ws.on_upgrade(socket_handlers::socket_handler));
+
+            let file_route = warp::path("ws")
+                .and(warp::path("file"))
+                .and(warp::ws())
+                .map(|ws: warp::ws::Ws| ws.on_upgrade(socket_handlers::file_handler));
 
             #[cfg(feature = "frontend")]
             let main_route = warp::any().map(|| {
@@ -115,7 +120,7 @@ fn main() {
                 .or(main_route)
                 .with(warp::compression::gzip());
 
-            let socket_routes = terminal_route.or(socket_route);
+            let socket_routes = terminal_route.or(socket_route).or(file_route);
 
             let routes = socket_routes
                 .or(login_route)
